+++ packages/adminui-vue/components/UTableEntity/filter-templates/date/range.vue
@@ -33,24 +33,38 @@ export default {
       value: null,
       pickerOptions: {
         shortcuts: [{
+          text: this.$ut('today'),
+          onClick: (picker) => {
+            const end = this.$UB.truncTimeToUtcNull(new Date())
+            const start = this.$UB.truncTimeToUtcNull(new Date())
+            picker.$emit('pick', [start, end])
+          }
+        }, {
+          text: this.$ut('yesterday'),
+          onClick: (picker) => {
+            const end = this.$UB.truncTimeToUtcNull(this.$moment().subtract(1, 'day').toDate())
+            const start = this.$UB.truncTimeToUtcNull(this.$moment().subtract(1, 'day').toDate())
+            picker.$emit('pick', [start, end])
+          }
+        }, {
           text: this.$ut('lastMonth'),
           onClick: (picker) => {
-            const end = new Date()
-            const start = this.$moment().subtract(1, 'month').toDate()
+            const end = this.$UB.truncTimeToUtcNull(new Date())
+            const start = this.$UB.truncTimeToUtcNull(this.$moment().subtract(1, 'month').toDate())
             picker.$emit('pick', [start, end])
           }
         }, {
           text: this.$ut('lastQuarter'),
           onClick: (picker) => {
-            const end = new Date()
-            const start = this.$moment().subtract(3, 'month').toDate()
+            const end = this.$UB.truncTimeToUtcNull(new Date())
+            const start = this.$UB.truncTimeToUtcNull(this.$moment().subtract(3, 'month').toDate())
             picker.$emit('pick', [start, end])
           }
         }, {
           text: this.$ut('last6Month'),
           onClick: (picker) => {
-            const end = new Date()
-            const start = this.$moment().subtract(6, 'month').toDate()
+            const end = this.$UB.truncTimeToUtcNull(new Date())
+            const start = this.$UB.truncTimeToUtcNull(this.$moment().subtract(6, 'month').toDate())
             picker.$emit('pick', [start, end])
           }
         }]